+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add `Connection.connect` and `Connection.connectOrFail` functions
 - Change inner type of `Client.Connected` to the `IGraphClient`, so the Client.Connected could be both `BoltGraphClient` or `GraphClient`
 - [**BC**] Support Neo4j Server `4.x` only
+- [**BC**] Rename `CypherQuery.toCypherResult` to `CypherQuery.toCypher` and remove a previous `CypherQuery.toCypher` function
 
 ## 2.0.0 - 2020-11-23
 - [**BC**] Use .netcore 5.0